// src/types/enums/inputTypes.enum.ts
export enum EInputTypeKeys {
  Name = "name",
  Phone = "phone",
  Email = "email",
  Address = "address",
  CreationDate = "creationDate",

  CreationDateFrom = "creationDateFrom",
  CreationDateTo = "creationDateTo",

  Diameter = "diameter",
  Material = "material",
  Price = "price",
  PurchasePrice = "purchasePrice",
  Quantity = "quantity",
}

export enum EInputTypeTitles {
  Name = "Имя",
  Phone = "Телефон",
  Email = "Email",
  Address = "Адрес доставки",
  CreationDate = "Дата регистрации",

  CreationDateFrom = "Дата регистрации c",
  CreationDateTo = "Дата регистрации до",

  Diameter = "Диаметр",
  Material = "Материал",
  Price = "Стоимость",
  PurchasePrice = "Цена закупки",
  Quantity = "Количество",
}